+++ packages/frontend/src/routes/Join/index.tsx
@@ -1,35 +1,6 @@
-import { User } from '@my-task/common';
-import { FormEventHandler, useRef, useState } from 'react';
-import { postJoin } from '~/api';
-import './index.css';
+import { RouteObject } from 'react-router-dom';
+import Join from './page';
 
-const Join = () => {
-  const [statusText, setStatusText] = useState('');
-  const emailInputRef = useRef<HTMLInputElement>(null);
-  const joinMutation = postJoin({
-    onSuccess: (data: User) => setStatusText(`User(${data.email}) has successfully joined!`),
-    onError: (err: { errorMessage: string; status: number }) => setStatusText(err.errorMessage),
-  });
+const joinRouteObject: RouteObject = { path: 'join', element: <Join /> };
 
-  const onSubmit: FormEventHandler = (e) => {
-    e.preventDefault();
-    if (!emailInputRef.current) return;
-
-    joinMutation.mutate({
-      email: emailInputRef.current.value,
-    });
-  };
-
-  return (
-    <form onSubmit={onSubmit}>
-      <div>
-        <label htmlFor="email">E-Mail</label>
-        <input id="email" type="email" ref={emailInputRef} />
-      </div>
-      {statusText && <div>{statusText}</div>}
-      <button type="submit">Sign Up</button>
-    </form>
-  );
-};
-
-export default Join;
+export default joinRouteObject;
