+++ packages/frontend/src/routes/Join/style.css
