+++ packages/frontend/src/routes/Root/index.tsx
@@ -1,37 +1,6 @@
-import { useState } from 'react';
-import { getConnectionTest } from '~/api';
-import reactLogo from '~/assets/react.svg';
-import viteLogo from '~/assets/vite.svg';
-import './index.css';
+import { RouteObject } from 'react-router-dom';
+import Root from './page';
 
-function Root() {
-  const [count, setCount] = useState(0);
-  const connection = getConnectionTest();
+const rootRouteObject: RouteObject = { path: '', element: <Root /> };
 
-  if (connection.data) console.log(connection.data);
-
-  return (
-    <div className="App">
-      <div>
-        <a href="https://vitejs.dev" target="_blank" rel="noreferrer">
-          <img src={viteLogo} className="logo" alt="Vite logo" />
-        </a>
-        <a href="https://reactjs.org" target="_blank" rel="noreferrer">
-          <img src={reactLogo} className="logo react" alt="React logo" />
-        </a>
-      </div>
-      <h1>Vite + React</h1>
-      <div className="card">
-        <button type="button" onClick={() => setCount((_count) => _count + 1)}>
-          count is {count}
-        </button>
-        <p>
-          Edit <code>src/App.tsx</code> and save to test HMR
-        </p>
-      </div>
-      <p className="read-the-docs">Click on the Vite and React logos to learn more</p>
-    </div>
-  );
-}
-
-export default Root;
+export default rootRouteObject;
