+++ packages/frontend/src/routes/Root/page.test.tsx
@@ -3,7 +3,7 @@ import { render, screen } from '@testing-library/react';
 import { describe, expect, it } from 'vitest';
 import { BE_ORIGIN } from '~/constants';
 import { server } from '~/mock';
-import Root from './index';
+import Root from './page';
 
 describe('App', () => {
   beforeAll(() => server.listen({ onUnhandledRequest: 'error' }));
