+++ packages/frontend/src/routes/index.ts
@@ -1,2 +1,7 @@
-import Root from './Root';
-export { Root };
+import { createBrowserRouter } from 'react-router-dom';
+import joinRouteObject from './Join';
+import rootRouteObject from './Root';
+
+const router = createBrowserRouter([rootRouteObject, joinRouteObject]);
+
+export default router;
